Another long --> size_t change, this time in the getPixelSet function for Object3D that is only used in askapsoft
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1002 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/PixelMap/Object3D.cc
@@ -322,7 +322,7 @@ namespace PixelInfo
 
   //--------------------------------------------------------------------
 
-  std::vector<Voxel> Object3D::getPixelSet(float *array, long *dim)
+  std::vector<Voxel> Object3D::getPixelSet(float *array, size_t *dim)
   {
     /// @details Returns a vector of Voxels with the flux values for each voxel
     /// taken from the array provided. No check is made as to whether
